SwipeDelegate: ensure background and contentItem are resized
When the geometry of a control changes, this code is called:

void QQuickControl::geometryChange(const QRectF &newGeometry, const QRectF &oldGeometry)
{
    Q_D(QQuickControl);
    QQuickItem::geometryChange(newGeometry, oldGeometry);
    d->resizeBackground();
    d->resizeContent();
    if (!qFuzzyCompare(newGeometry.width(), oldGeometry.width()))
        emit availableWidthChanged();
    if (!qFuzzyCompare(newGeometry.height(), oldGeometry.height()))
        emit availableHeightChanged();
}

SwipeDelegate works by moving the background and contentItem together
when it is swiped to expose the various delegates. Because this
involves setting the position of the background, the check for
background's x position being 0 in
QQuickControlPrivate::resizeBackground() would fail and the background
would not be resized at all. Fix this by making resizeBackground()
virtual and don't check the x when checking if we should set the width.

Similarly, in QQuickSwipeDelegatePrivate::resizeContent(), we should
set the contentItem's width instead of just repositioning and resizing
it vertically.

Fixes: QTBUG-85770
Pick-to: 5.15 6.0
Change-Id: I36684bf2797719db87fe93063cc7685efe594eea
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 83f49d3f4c0041bf2c74ff1ff2eb03c5d37aa1f4)