Reset the opacity and scale properties after the exit transition
By resetting the opacity and scale properties after an exit transition
we are ensuring that it does not lose the original values that the user
may have set.

[ChangeLog][Important Behavior Changes][Popup] After the exit
transition is finished, then the opacity and scale properties will
be reset to their values before the enter transition is started.

Pick-to: 5.15
Fixes: QTBUG-87283
Change-Id: I2b192c96eaea2906d968341255e80cd19be177e6
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit df33c79fb6579f94498164531777320e79c36fd5)